+++ js/mod.js
@@ -1,30 +1,30 @@
 let modInfo = {
 	name: "Prestige Tree Rewritten",
 	id: "ptr",
-	author: "Jacorb (Despacit helped port)",
+	author: "Jacorb",
 	pointsName: "points",
 	discordName: "PT Rewritten Server",
 	discordLink: "https://discord.gg/TFCHJJT",
 	changelogLink: "https://github.com/AbitofTetration/Prestige-Tree-Rewritten/blob/master/changelog.md",
     offlineLimit: 1,  // In hours
     initialStartPoints: new Decimal(10), // Used for hard resets and new players
-	endgame: new Decimal("e1.25e10"),
-	// specialEndgameText: "v1.1 Beta 21 Endgame: e1.25e10 Points & 100 Phantom Souls",
+	endgame: new Decimal("e1e15"),
+	// specialEndgameText: "v1.2 Beta 28 Endgame: e1e15 Points",
 }
 
 // Set your version in num and name
 let VERSION = {
-	num: "1.1",
-	// beta: 21,
+	num: "1.2",
+	// beta: 28,
 	// patch: 1,
-	name: "Hyperium Nebulae",
+	name: "Mechanical Mastery",
 }
 
 // If you add new functions anywhere inside of a layer, and those functions have an effect when called, add them here.
 // (The ones here are examples, all official functions are already taken care of)
-var doNotCallTheseFunctionsEveryTick = ["doReset", "buy", "onPurchase", "blowUpEverything", "castAllSpells"]
+var doNotCallTheseFunctionsEveryTick = ["doReset", "buy", "onPurchase", "blowUpEverything", "castAllSpells", "completeInBulk", "startMastery", "completeMastery"]
 
-var alwaysKeepTheseVariables = ["primeMiles", "auto", "autoExt", "autoBld", "autoW", "keepPosNeg", "distrAll", "spellInput", "pseudoUpgs"]
+var alwaysKeepTheseVariables = ["primeMiles", "auto", "autoExt", "autoBld", "autoW", "autoGhost", "keepPosNeg", "distrAll", "spellInput", "pseudoUpgs", "maxToggle"]
 
 function getStartPoints(){
     return new Decimal(modInfo.initialStartPoints)
@@ -45,6 +45,7 @@ function getPointGen() {
 	if (hasUpgrade("p", 13)) gain = gain.times(upgradeEffect("p", 13));
 	if (hasUpgrade("p", 22)) gain = gain.times(upgradeEffect("p", 22));
 	if (hasUpgrade("b", 14) && player.i.buyables[12].gte(1)) gain = gain.times(upgradeEffect("b", 11))
+	if (((Array.isArray(tmp.ma.mastered))?tmp.ma.mastered.includes("e"):false) && hasUpgrade("e", 12)) gain = gain.times(upgradeEffect("e", 12))
 	if (hasAchievement("a", 21)) gain = gain.times(1.1);
 	if (hasAchievement("a", 31)) gain = gain.times(1.5);
 	if (inChallenge("h", 22)) return gain.times(player.s.unlocked?buyableEffect("s", 11):1).root(inChallenge("h", 31)?tmp.h.pointRoot31:1);
@@ -57,9 +58,15 @@ function getPointGen() {
 	if (player.q.unlocked) gain = gain.times(tmp.q.enEff);
 	
 	if (inChallenge("h", 31)) gain = gain.root(tmp.h.pointRoot31);
-	if (hasUpgrade("ss", 43)) gain = gain.pow(gain.lt("e1e6")?1.1:1.01);
+	if (hasUpgrade("ss", 43)) gain = gain.pow(gain.lt(tmp.ss.upgrades[43].endpoint)?1.1:1.01);
 	if (hasUpgrade("hn", 31)) gain = gain.pow(1.05);
 	return gain
+}
+
+function getRow1to6Speed() {
+	let speed = new Decimal(1);
+	if ((Array.isArray(tmp.ma.mastered))?tmp.ma.mastered.includes("t"):false) speed = speed.times(tmp.t.effect2)
+	return speed;
 }
 
 // You can add non-layer related variables that should to into "player" and be saved here, along with default values